% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2day.R
\name{date2day_fun}
\alias{date2day_fun}
\title{Date2Day function}
\usage{
date2day_fun(year, month, day)
}
\arguments{
\item{year}{Enter the year in an int format, eg. 1989}

\item{month}{Enter the month in an int format, eg. 8}

\item{day}{Enter the day in an int format, eg. 25}
}
\description{
This function allows you to calculate the day based on date.
The formula used in the function is kim Larsson calculation formula.
}
\examples{
date2day_fun(2018, 5, 28)

}
\keyword{Date2Day}
